move `book` and `label` class into classes folder
