+++ classes/label.rb
